fix load error for terminal-table - add board class in board folder
