+++ Gemfile
@@ -1,3 +1,3 @@
 gem 'rubocop'
 source 'https://rubygems.org'
-gem 'terminal-table', '~> 1.8'
+gem 'terminal-table', '~> 1.7', '>= 1.7.3'
